rename header and footer
